Criteria for evaluating places and rating criteria have been added.

+++ Domain/Interfaces/IPlaceRepository.cs
@@ -5,9 +5,9 @@ namespace JoyTopBackend.Domain.Interfaces;
 public interface IPlaceRepository
 {
     Task<IEnumerable<Place>> GetAllAsync();
-    Task<Place?> GetByIdAsync(int id);
+    Task<Place?> GetByIdAsync(long id);
     Task<IEnumerable<Place>> GetByCategoryAsync(string category);
     Task<Place> AddAsync(Place place);
     Task UpdateAsync(Place place);
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
